pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "flask>=3.1.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "ta>=0.11.0",
    "telegram>=0.0.1",
    "yfinance>=0.2.66",
]
